requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]==3.12
certifi
scikit-learn
scipy
dill
pyaml
# -e .